updated apis of course

+++ backend/src/modules/configurations/course-specialization/extras/association.service.ts
@@ -21,33 +21,38 @@ export class AssociationService extends GenericService<
   constructor(
     @InjectModel(Associations)
     private associationModel: typeof Associations,
-
     private reqParams: RequestParamsService,
   ) {
     super(associationModel, reqParams);
   }
 
-  async updateAssociationImage(file: Express.Multer.File, id: string) {
-    const association = await this.getOne<Associations>(id);
-    const defaultImagePath = 'backend/src/public/media/default.png';
-    const filePath = join(
-      __dirname,
-      '../../../../',
-      '/src/public/' + association.image,
-    );
+  async create<Model extends {} = any>(dto: AssociationsDTO): Promise<Model> {
+    return super.create(dto);
+  }
 
-    if (fs.existsSync(filePath)) {
-      unlink(filePath, (err) => {
-        if (err) {
-          throw new InternalServerErrorException(err);
-        }
-        console.log('file deleted...');
+  async updateAssociationImage(file: Express.Multer.File, id: string) {
+    if (file) {
+      const association = await this.getOne<Associations>(id);
+      const defaultImagePath = 'backend/src/public/media/default.png';
+      const filePath = join(
+        __dirname,
+        '../../../../',
+        '/src/public/' + association.image,
+      );
+
+      if (fs.existsSync(filePath)) {
+        unlink(filePath, (err) => {
+          if (err) {
+            throw new InternalServerErrorException(err);
+          }
+          console.log('file deleted...');
+        });
+      }
+
+      await association.update({
+        image: '/media/associations/' + file.filename,
       });
     }
-
-    await association.update({
-      image: '/media/association/' + file.filename,
-    });
     return 'Association Image Uploaded Successfully';
   }
 }
